refactor:#39 header 로고 표시 유무 props 추가
